add 凸予定を全て削除するコマンドの実装
